requirements.txt
django
django_rest
psycopg2-binary
djangorestframework
django-environ
Pillow
django_extensions
djoser
djangorestframework-simplejwt
redis
celery
drf-spectacular
